+++ identity_ecdsa_verifier/src/lib.rs
@@ -3,14 +3,14 @@
 
 #![doc = include_str!("./../README.md")]
 #![warn(
-    rust_2018_idioms,
-    unreachable_pub,
-    missing_docs,
-    rustdoc::missing_crate_level_docs,
-    rustdoc::broken_intra_doc_links,
-    rustdoc::private_intra_doc_links,
-    rustdoc::private_doc_tests,
-    clippy::missing_safety_doc
+  rust_2018_idioms,
+  unreachable_pub,
+  missing_docs,
+  rustdoc::missing_crate_level_docs,
+  rustdoc::broken_intra_doc_links,
+  rustdoc::private_intra_doc_links,
+  rustdoc::private_doc_tests,
+  clippy::missing_safety_doc
 )]
 
 mod ecdsa_jws_verifier;
