+++ identity_ecdsa_verifier/src/tests/mod.rs
@@ -1,0 +1,5 @@
+// Copyright 2020-2024 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+mod secp256;
+mod secp256k;
